feat(ui): move new-request intercept to useOverlayIntercepts
